requirements.txt
# GPU 數據收集與視覺化專案依賴套件
# 
# 如需使用視覺化功能，請安裝 visualization/ 資料夾內的依賴：
# pip install -r visualization/requirements.txt
#
# 或使用執行腳本自動安裝：
# ./run_gpu_visualization.sh

# 基本數據處理依賴（用於數據收集腳本）
pandas>=1.3.0

# 環境變數載入
python-dotenv>=1.0.0

# HTTP 客戶端 (async)
aiohttp>=3.8.0

# CLI 框架
click>=8.1.0

# 視覺化依賴請參考 visualization/requirements.txt
